+++ sac/src/Page/MyPage.js
@@ -1,17 +1,17 @@
 import React from "react";
 import '../styles/MyPage.css';
 import Header from "../components/Header/Header";
-import MyHead from "../components/MyPage/MyHead";
+import MyNavi from "../components/MyPage/MyNavi";
 
 function MyPage() {
 
   return (
     <div className="App">
-      <div className="header">
+      <div className="navi">
         <Header />
       </div>
       <div className="body">
-        <MyHead />
+        <MyNavi />
       </div>
     </div>
   );
