+++ sac/src/components/Navi/LeftNavi/LeftNavi.js
@@ -1,5 +1,6 @@
 import React from "react";
 import "../../../styles/LeftNavi.css";
+import TopicItem from "../../commons/TopicItem";
 
 const topics = ["프론트엔드", "백엔드", "맛집", "전자기기", "축구", "영화", "건강", "문화생활", "뷰티", "음주", "연애", "예능", "드라마"];
 
@@ -25,47 +26,5 @@ function LeftNavi() {
     );
 }
 
-//토픽을 나타내는 컴포넌트
-function TopicItem({ topic }) {
-    let icon = "🌱";
-    if (topic === "프론트엔드") {
-        icon = "🌈";
-    } else if (topic === "백엔드") {
-        icon = "💻";
-    } else if (topic === "맛집") {
-        icon = "🍔";
-    } else if (topic === "전자기기") {
-        icon = "📱";
-    } else if (topic === "축구") {
-        icon = "⚽️";
-    } else if (topic === "영화") {
-        icon = "🎬";
-    } else if (topic === "건강") {
-        icon = "🏋️‍♀️";
-    } else if (topic === "문화생활") {
-        icon = "🎨";
-    } else if (topic === "뷰티") {
-        icon = "💄";
-    } else if (topic === "음주") {
-        icon = "🍺";
-    }
-    else if (topic === "연애") {
-        icon = "💕";
-    }
-    else if (topic === "예능") {
-        icon = "🤣";
-    }
-    else if (topic === "드라마") {
-        icon = "📺";
-    }
-
-    return (
-        
-        <div className="topic-item">
-            <p className="topic-item-icon">{icon}</p>
-            <p className="topic-item-text">{topic}</p>
-        </div>
-    );
-}
 
 export default LeftNavi;
